greatly simplifying project subscriptions
