+++ packages/frontend-2/pages/projects/[id]/models/index.vue
@@ -26,24 +26,12 @@
 <script setup lang="ts">
 import { useQuery } from '@vue/apollo-composable'
 import { SourceAppDefinition } from '@speckle/shared'
-import { ToastNotificationType, useGlobalToast } from '~~/lib/common/composables/toast'
-import {
-  FormUsersSelectItemFragment,
-  ProjectModelsUpdatedMessageType,
-  ProjectUpdatedMessageType
-} from '~~/lib/common/generated/gql/graphql'
-import { useNavigateToHome } from '~~/lib/common/helpers/route'
-import {
-  useEvictProjectModelFields,
-  useProjectModelUpdateTracking
-} from '~~/lib/projects/composables/modelManagement'
-import { useProjectUpdateTracking } from '~~/lib/projects/composables/projectManagement'
-import {
-  useProjectPendingVersionUpdateTracking,
-  useProjectVersionUpdateTracking
-} from '~~/lib/projects/composables/versionManagement'
+import { FormUsersSelectItemFragment } from '~~/lib/common/generated/gql/graphql'
 import { projectModelsPageQuery } from '~~/lib/projects/graphql/queries'
-import { useProjectPageItemViewType } from '~~/lib/projects/composables/layout'
+import {
+  useGeneralProjectPageUpdateTracking,
+  useProjectPageItemViewType
+} from '~~/lib/projects/composables/projectPages'
 
 definePageMeta({
   middleware: ['require-valid-project']
@@ -51,9 +39,6 @@ definePageMeta({
 
 const route = useRoute()
 const projectId = computed(() => route.params.id as string)
-const evictProjectModels = useEvictProjectModelFields()
-const { triggerNotification } = useGlobalToast()
-const goHome = useNavigateToHome()
 
 const selectedMembers = ref([] as FormUsersSelectItemFragment[])
 const selectedApps = ref([] as SourceAppDefinition[])
@@ -66,30 +51,7 @@ const { result } = useQuery(projectModelsPageQuery, () => ({
 }))
 
 const project = computed(() => result.value?.project)
-
-// TODO: Maybe move these to a parent page that is the same for all projects subpages?
-// Subscriptions for tracking updates to project, models, versions
-useProjectVersionUpdateTracking(projectId)
-useProjectPendingVersionUpdateTracking(projectId)
-useProjectModelUpdateTracking(projectId, (event) => {
-  // If creation, refresh all project's model fields
-  if (event.type === ProjectModelsUpdatedMessageType.Created) {
-    evictProjectModels(projectId.value)
-  }
-})
-useProjectUpdateTracking(projectId, (event) => {
-  const isDeleted = event.type === ProjectUpdatedMessageType.Deleted
-
-  if (isDeleted) {
-    goHome()
-  }
-
-  triggerNotification({
-    type: ToastNotificationType.Info,
-    title: isDeleted ? 'Project deleted' : 'Project updated',
-    description: isDeleted ? 'Redirecting to home' : undefined
-  })
-})
+useGeneralProjectPageUpdateTracking({ projectId })
 
 const clearSearch = () => {
   search.value = ''
